+++ age/src/error.rs
@@ -3,9 +3,30 @@
 use std::fmt;
 use std::io;
 
+/// The various errors that can be returned during the encryption process.
+#[derive(Debug)]
+pub enum EncryptError {
+    /// An I/O error occurred during decryption.
+    Io(io::Error),
+}
+
+impl From<io::Error> for EncryptError {
+    fn from(e: io::Error) -> Self {
+        EncryptError::Io(e)
+    }
+}
+
+impl fmt::Display for EncryptError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            EncryptError::Io(e) => e.fmt(f),
+        }
+    }
+}
+
 /// The various errors that can be returned during the decryption process.
 #[derive(Debug)]
-pub enum Error {
+pub enum DecryptError {
     /// The age file failed to decrypt.
     DecryptionFailed,
     /// The age file used an excessive work factor for passphrase encryption.
@@ -29,11 +50,11 @@ pub enum Error {
     UnknownFormat,
 }
 
-impl fmt::Display for Error {
+impl fmt::Display for DecryptError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Error::DecryptionFailed => write!(f, "Decryption failed"),
-            Error::ExcessiveWork { required, target } => {
+            DecryptError::DecryptionFailed => write!(f, "Decryption failed"),
+            DecryptError::ExcessiveWork { required, target } => {
                 writeln!(f, "Excessive work parameter for passphrase.")?;
                 write!(
                     f,
@@ -41,12 +62,12 @@ impl fmt::Display for Error {
                     1 << (required - target)
                 )
             }
-            Error::InvalidHeader => write!(f, "Header is invalid"),
-            Error::InvalidMac => write!(f, "Header MAC is invalid"),
-            Error::Io(e) => e.fmt(f),
-            Error::KeyDecryptionFailed => write!(f, "Failed to decrypt an encrypted key"),
-            Error::NoMatchingKeys => write!(f, "No matching keys found"),
-            Error::UnknownFormat => {
+            DecryptError::InvalidHeader => write!(f, "Header is invalid"),
+            DecryptError::InvalidMac => write!(f, "Header MAC is invalid"),
+            DecryptError::Io(e) => e.fmt(f),
+            DecryptError::KeyDecryptionFailed => write!(f, "Failed to decrypt an encrypted key"),
+            DecryptError::NoMatchingKeys => write!(f, "No matching keys found"),
+            DecryptError::UnknownFormat => {
                 writeln!(f, "Unknown age format.")?;
                 write!(f, "Have you tried upgrading to the latest version?")
             }
@@ -54,35 +75,35 @@ impl fmt::Display for Error {
     }
 }
 
-impl From<chacha20poly1305::aead::Error> for Error {
+impl From<chacha20poly1305::aead::Error> for DecryptError {
     fn from(_: chacha20poly1305::aead::Error) -> Self {
-        Error::DecryptionFailed
+        DecryptError::DecryptionFailed
     }
 }
 
-impl From<io::Error> for Error {
+impl From<io::Error> for DecryptError {
     fn from(e: io::Error) -> Self {
-        Error::Io(e)
+        DecryptError::Io(e)
     }
 }
 
-impl From<hmac::crypto_mac::MacError> for Error {
+impl From<hmac::crypto_mac::MacError> for DecryptError {
     fn from(_: hmac::crypto_mac::MacError) -> Self {
-        Error::InvalidMac
+        DecryptError::InvalidMac
     }
 }
 
 #[cfg(feature = "ssh")]
-impl From<rsa::errors::Error> for Error {
+impl From<rsa::errors::Error> for DecryptError {
     fn from(_: rsa::errors::Error) -> Self {
-        Error::DecryptionFailed
+        DecryptError::DecryptionFailed
     }
 }
 
-impl std::error::Error for Error {
+impl std::error::Error for DecryptError {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
         match self {
-            Error::Io(inner) => Some(inner),
+            DecryptError::Io(inner) => Some(inner),
             _ => None,
         }
     }