+++ rage/src/bin/rage/error.rs
@@ -14,6 +14,14 @@ pub(crate) enum EncryptError {
     UnknownAlias(String),
 }
 
+impl From<age::EncryptError> for EncryptError {
+    fn from(e: age::EncryptError) -> Self {
+        match e {
+            age::EncryptError::Io(e) => EncryptError::Io(e),
+        }
+    }
+}
+
 impl From<io::Error> for EncryptError {
     fn from(e: io::Error) -> Self {
         EncryptError::Io(e)
@@ -65,7 +73,7 @@ impl fmt::Display for EncryptError {
 }
 
 pub(crate) enum DecryptError {
-    Age(age::Error),
+    Age(age::DecryptError),
     ArmorFlag,
     IdentityNotFound(String),
     Io(io::Error),
@@ -79,8 +87,8 @@ pub(crate) enum DecryptError {
     UnsupportedKey(String, age::ssh::UnsupportedKey),
 }
 
-impl From<age::Error> for DecryptError {
-    fn from(e: age::Error) -> Self {
+impl From<age::DecryptError> for DecryptError {
+    fn from(e: age::DecryptError) -> Self {
         DecryptError::Age(e)
     }
 }
@@ -95,7 +103,7 @@ impl fmt::Display for DecryptError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             DecryptError::Age(e) => match e {
-                age::Error::ExcessiveWork { required, .. } => {
+                age::DecryptError::ExcessiveWork { required, .. } => {
                     writeln!(f, "{}", e)?;
                     write!(f, "To decrypt, retry with --max-work-factor {}", required)
                 }